requirements.txt
discord.py
pillow
pet-pet-gif
pilmoji
python-urbandict
psutil
pyyaml
sqlitecloud
praw
plyvel
beautifulsoup4
aiohttp